+++ src/asp/Camera/BundleAdjustIsis.h
@@ -56,7 +56,6 @@ struct IsisCnetData {
 // Load an ISIS cnet file and copy it to an ASP control network. The ISIS cnet
 // will be used when saving the updated cnet.  
 void loadIsisCnet(std::string const& isisCnetFile, 
-                  std::string const& outputPrefix, 
                   std::vector<std::string> const& image_files,
                   // Outputs
                   vw::ba::ControlNetwork& cnet,
@@ -64,9 +63,9 @@ void loadIsisCnet(std::string const& isisCnetFile,
 
 // Update an ISIS cnet with the latest info on triangulated points and outliers,
 // and write it to disk at <outputPrefix>.net.
-void saveUpdatedIsisCnet(std::string const& outputPrefix, 
-                  asp::BAParams const& param_storage,
-                  IsisCnetData & isisCnetData);
+void saveUpdatedIsisCnet(std::string const& outputPrefix,
+                         asp::BAParams const& param_storage,
+                         IsisCnetData & isisCnetData);
 
 // Create and save an ISIS cnet from a given control network and latest param
 // values.